% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEanalysis-package.R
\docType{package}
\name{DEanalysis-package}
\alias{DEanalysis-package}
\alias{DEanalysis}
\title{Differential expression analysis and performance test on differentail expression methods}
\description{
RNAseq data simulation, differential expression analysis and
performance comparison of differential expression methods
}
\details{
This package provides extensive functionality for comparing
results obtained by different methods for differential expression analysis
of RNAseq data. It also contains functions for simulating count data and
interfaces to several packages for performing the differential expression analysis.
}
\author{
BuKyung Baik
}
\keyword{package}
