% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceplot.R
\name{fpc_performance_plot}
\alias{fpc_performance_plot}
\title{make synthetic data False Positive count plot function}
\usage{
fpc_performance_plot(working.dir, figure.dir, simul.data, rep, nsample,
  disp.Type, modes, AnalysisMethods)
}
\arguments{
\item{working.dir}{Input file location}

\item{figure.dir}{Figure save location}

\item{simul.data}{Type of dataset (e.g. KIRC, Bottomly, mBdK and mKdB)}

\item{rep}{Number of replication each test contain.}

\item{nsample}{Number of samples. Input as a numeric vector.}

\item{disp.Type}{How is the dispersion assumed for each condition. Possible values are 'same' and 'different'.}

\item{modes}{Test conditions we used for simulation data generation. Input as a character vector.
"D" for basic simulation (not adding outliers).
"R" for adding 5% of random outlier.
"OS"for adding outlier sample to each sample group.}

\item{AnalysisMethods}{DEmethods used for figures. Input as character vectors
(e.g. 'edgeR','edgeR.ql','edgeR.rb','DESeq.pd','DESeq2','voom.tmm','voom.qn','voom.sw','ROTS','BaySeq','PoissonSeq','SAMseq')}
}
\description{
make synthetic data False Positive count plot function
}
