% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationDataGeneration.R
\name{generateDatasetParameter}
\alias{generateDatasetParameter}
\title{Generate List containing estimated mean and dispersion parameters and filtered count from original count dataset.}
\usage{
generateDatasetParameter()
}
\description{
Generate List containing estimated mean and dispersion parameters and filtered count from original count dataset.
}
