% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEAnalysis.R
\name{GenerateRealSimulation}
\alias{GenerateRealSimulation}
\title{Generation of Real simulation files}
\usage{
GenerateRealSimulation(working.dir, fpc = FALSE, data.types, rep,
  nsample)
}
\arguments{
\item{working.dir}{Input file location}

\item{fpc}{A logical indicating whether simulation data is made from single sample condition(e.g. normal) to calculate False positive counts. Possible values are TRUE or FALSE.}

\item{data.types}{A vector indicating types of dataset (e.g. data.types = c(KIRC, Bottomly))}

\item{rep}{An integer specifying iterations each test perform.}

\item{nsample}{A vector indicating number of samples. Possible values can be 1~72 for KIRC, 1~10 for Bottomly and 1~5 for SEQC, maximum values decreased to 36 for KIRC and 5 for Bottomly if fpc is true.}
}
\description{
Generation of Real simulation files
}
