% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEAnalysis.R
\name{GenerateSyntheticSimulation}
\alias{GenerateSyntheticSimulation}
\title{Generation of Synthetic simulation files}
\usage{
GenerateSyntheticSimulation(working.dir, data.types, fixedfold = FALSE,
  rep, nsample, nvar, nDE, fraction.upregulated = 0.5, disp.Types, modes)
}
\arguments{
\item{working.dir}{Input file location}

\item{data.types}{A vector parameter indicating types of dataset (e.g. data.type = c(KIRC, Bottomly, mBdK and mKdB))}

\item{fixedfold}{A logical indicating whether simulation data is made from fixed fold to imitate SEQC counts data. Possible values are TRUE or FALSE. If fixedfold is TRUE, fraction of upregulated genes is automatically fixed to 0.67.}

\item{rep}{An integer specifying iterations each test perform.}

\item{nsample}{A vector indicating number of samples.}

\item{nvar}{An integer indicating the number of total gene in the synthetic data.}

\item{nDE}{A vector indicating number of generated DE genes in the synthetic data.}

\item{fraction.upregulated}{A vector specifying proportion of upregulated DE genes in the synthetic data. Default value is 0.5. (e.g. fraction.upregulated = c(0.5, 0.7 and 0.9)) This vector is available only if fixed fold is FALSE.}

\item{disp.Types}{A vector indicating how is the dispersion parameter assumed to be for each condition to make a synthetic data. Possible values are 'same' and 'different'.}

\item{modes}{A vector specifying test conditions we used for simulation data generation.
"D" for basic simulation (not adding outliers).
"R" for adding 5% of random outlier.
"OS"for adding outlier sample to each sample group.}
}
\description{
Generation of Synthetic simulation files
}
