% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationDataGeneration.R
\name{getDisp}
\alias{getDisp}
\title{Get random dispersion value based on the read count size (mean).}
\usage{
getDisp(mean, mean.condition, disp.condition)
}
\arguments{
\item{mean}{read count size}

\item{mean.condition}{read count vector of a certain condition(normal/cancer)}

\item{disp.condition}{dispersion vector of a certain condition(normal/cancer)}
}
\description{
Get random dispersion value based on the read count size (mean).
}
