% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceplot.R
\name{performance_plot}
\alias{performance_plot}
\title{make synthetic data plot function}
\usage{
performance_plot(working.dir, figure.dir, simul.data, fixedfold = FALSE,
  rep, nsample, nvar, nDE, fraction.upregulated, disp.Type, mode, rowType,
  AnalysisMethods)
}
\arguments{
\item{working.dir}{Input file location}

\item{figure.dir}{Figure save location}

\item{simul.data}{Type of dataset (e.g. KIRC, Bottomly, mBdK and mKdB)}

\item{fixedfold}{A logical indicating whether simulation data is made from fixed fold to imitate SEQC counts data. Possible values are TRUE or FALSE. If fixedfold is TRUE, fraction.upregulated is automatically fixed to 0.67.}

\item{rep}{Number of replication each test contain.}

\item{nsample}{Number of samples. Input as a numeric vector.}

\item{nvar}{Number of total gene making synthetic data.}

\item{nDE}{Number of generated DE genes in the synthetic data.}

\item{fraction.upregulated}{proportion of upregulated DE genes in the synthetic data (e.g. 0.5, 0.7 and 0.9)}

\item{disp.Type}{How is the dispersion assumed for each condition. Possible values are 'same' and 'different'.}

\item{mode}{Test conditions we used for simulation data generation. Input as a character vector.
"D" for basic simulation (not adding outliers).
"R" for adding 5% of random outlier.
"OS" for adding outlier sample to each sample group.
"DL" for decreasing KIRC simulation dispersion 22.5 times (similar to SEQC data dispersion) to compare with SEQC data.}

\item{rowType}{Type of measures. Combination of AUC, TPR and trueFDR. (e.g. c('AUC','TPR'))}

\item{AnalysisMethods}{DEmethods used for figures. Input as character vectors
(e.g. 'edgeR','edgeR.ql','edgeR.rb','DESeq.pd','DESeq2','voom.tmm','voom.qn','voom.sw','ROTS','BaySeq','BaySeq.qn','PoissonSeq','SAMseq')}
}
\description{
make synthetic data plot function
}
