% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonSeq.createRmd.R
\name{PoissonSeq.createRmd}
\alias{PoissonSeq.createRmd}
\title{Generate a \code{.Rmd} file containing code to perform differential expression analysis with a Poisson log linear model}
\usage{
PoissonSeq.createRmd(data.path, result.path, codefile)
}
\arguments{
\item{data.path}{The path to a .rds file containing the \code{compData} object that will be used for the differential expression analysis.}

\item{result.path}{The path to the file where the result object will be saved.}

\item{codefile}{The path to the file where the code will be written.}
}
\value{
The function generates a \code{.Rmd} file containing the code for performing the differential expression analysis. This file can be executed using e.g. the \code{knitr} package.
}
\description{
A function to generate code that can be run to perform differential expression analysis of RNAseq data (comparing two conditions) using the PoissonSeq package. The code is written to a \code{.Rmd} file. This function is generally not called by the user, the main interface for performing differential expression analysis is the \code{\link{runDiffExp}} function.
}
\details{
For more information about the methods and the interpretation of the parameters, see the \code{PoissonSeq} package and the corresponding publications.
}
\examples{
tmpdir <- normalizePath(tempdir(), winslash = "/")
SyntheticDataSimulation(simul.data= 'KIRC', dataset = file.path(tmpdir, "mydata.rds"), n.var = 500,
                                    samples.per.cond = 3, n.diffexp = 100, dispType='same', mode='D',
                                    fraction.upregulated=0.5, dataset.parameters=generateDatasetParameter())
compcodeR::runDiffExp(data.file = file.path(tmpdir, "mydata.rds"), result.extent = "PoissonSeq",
           Rmdfunction = "PoissonSeq.createRmd",
           output.directory = tmpdir)
}
\references{
Li J, Witten DM, Johnstone I, Tibshirani R (2012). Normalization, testing, and false discovery rate estimation for RNA-sequencing data. Biostatistics 13(3): 523-38.
}
\author{
BuKyung Baik
}
