% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROTS.createRmd.R
\name{ROTS.createRmd}
\alias{ROTS.createRmd}
\title{Generate a \code{.Rmd} file containing code to perform differential expression analysis with Reproducibility-Optimized Test Statistic(ROTS)}
\usage{
ROTS.createRmd(data.path, result.path, codefile, normalize = TRUE,
  transformation = TRUE, B = 1000, K = NULL, log = FALSE)
}
\arguments{
\item{data.path}{The path to a .rds file containing the \code{compData} object that will be used for the differential expression analysis.}

\item{result.path}{The path to the file where the result object will be saved.}

\item{codefile}{The path to the file where the code will be written.}

\item{normalize}{A logical indicating whether to run ROTS with TMM normalized count or raw count data. Default value is True (TMM normalized).}

\item{transformation}{A logical indicating whether to run ROTS with voom transformation. Default value is True.}

\item{B}{An integer specifying the number of bootstrap and permutation resamplings (default 1000).}

\item{K}{An integer indicating the largest top list size considered. If no value is given, 1/4 of the features are used.}

\item{log}{A logical parameter indicating whether input data is log2 scaled. Choices are as for the method argument of ROTS function. This information is only used to calculate log fold change.}
}
\value{
The function generates a \code{.Rmd} file containing the code for performing the differential expression analysis. This file can be executed using e.g. the \code{knitr} package.
}
\description{
A function to generate code that can be run to perform differential expression analysis of RNAseq data (comparing two conditions) using the ROTS package.
}
\details{
For more information about the methods and the interpretation of the parameters, see the \code{ROTS} package and the corresponding publications.
}
\examples{
tmpdir <- normalizePath(tempdir(), winslash = "/")
SyntheticDataSimulation(simul.data= 'KIRC', dataset = file.path(tmpdir, "mydata.rds"), n.var = 500,
                                    samples.per.cond = 3, n.diffexp = 100, dispType='same', mode='D',
                                    fraction.upregulated=0.5, dataset.parameters=generateDatasetParameter())
compcodeR::runDiffExp(data.file = file.path(tmpdir, "mydata.rds"), result.extent = "ROTS",
           Rmdfunction = "ROTS.createRmd", transformation=TRUE, normalize = TRUE,
           output.directory = tmpdir, B = 1000,
           K = 500, log = FALSE)
}
\references{
L. L. Elo, S. Filen, R. Lahesmaa and T. Aittokallio: Reproducibility-optimized test statistic for ranking genes in microarray studies. IEEE/ACM Transactions on Computational Biology and Bioinformatics 5: 423–431, 2008.
}
\author{
BuKyung Baik
}
