% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceplot.R
\name{run_PCA}
\alias{run_PCA}
\title{run PCA}
\usage{
run_PCA(datatypes)
}
\arguments{
\item{datatypes}{A character parameter indicating the object dataset to be analyzed with PCA.}
}
\description{
Run PCA on two sample datasets, KIRC and Bottomly.
}
