% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEAnalysis.R
\name{runSimulationAnalysis}
\alias{runSimulationAnalysis}
\title{Run Analysismethods}
\usage{
runSimulationAnalysis(working.dir, output.dir, real = FALSE,
  fpc = FALSE, data.types, fixedfold = FALSE, rep, nsample, nDE,
  fraction.upregulated, disp.Types, modes, AnalysisMethods,
  para = list())
}
\arguments{
\item{working.dir}{Input file location}

\item{output.dir}{Result file location}

\item{real}{A logical indicating whether this analysis is for synthetic data or real data. Possible values are TRUE or FALSE.}

\item{fpc}{A logical indicating whether simulation data is made from a single sample group (e.g. normal) to calculate False positive counts. Possible values are TRUE or FALSE. Only used for real data analysis.}

\item{data.types}{A vector indicating types of dataset (e.g. data.types = c(KIRC, Bottomly, mBdK and mKdB))}

\item{fixedfold}{A logical indicating whether simulation data is made from fixed fold to imitate SEQC counts data. Possible values are TRUE or FALSE. Only applied if this simulation is synthetic data simulation.}

\item{rep}{An integer specifying iterations each test perform.}

\item{nsample}{A vector indicating number of samples.}

\item{nDE}{A vector indicating number of generated DE genes in the synthetic data.}

\item{fraction.upregulated}{A vector specifying proportion of upregulated DE genes in the synthetic data. Default value is 0.5. (e.g. fraction.upregulated = c(0.5, 0.7 and 0.9)) This vector is available only if fixed fold is FALSE.}

\item{disp.Types}{A vector indicating how is the dispersion parameter assumed to be for each condition to make a synthetic data. Possible values are 'same' and 'different'.}

\item{modes}{A vector specifying test conditions we used for simulation data generation.
"D" for basic simulation (not adding outliers).
"R" for adding 5% of random outlier.
"OS"for adding outlier sample to each sample group.
"DL" for decreasing KIRC simulation dispersion 22.5 times (similar to SEQC data dispersion) to compare with SEQC data.}

\item{AnalysisMethods}{A vector specifying DEmethods used for the analysis.
(e.g. 'edgeR','edgeR.ql','edgeR.rb','DESeq.pd','DESeq2','voom.tmm','voom.qn','voom.sw','ROTS','BaySeq','BaySeq.qn,'PoissonSeq','SAMseq')}

\item{para}{A list parameter consists of multiple lists corresponding each method. Each method list contains multiple parameters to run each DE analysis method.}
}
\description{
Run Analysismethods
}
