% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceplot.R
\name{select_color}
\alias{select_color}
\title{color selector}
\usage{
select_color(method_color)
}
\arguments{
\item{method_color}{A character parameter indicating the object method to be switched into its corresponding color in the performance plot.}
}
\description{
Take input methods and return pre-assigned color for figure.
}
