% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceplot.R
\name{select_tool}
\alias{select_tool}
\title{tool selector}
\usage{
select_tool(method_in)
}
\arguments{
\item{method_in}{A character parameter indicating the object method to be switched into its file extension.}
}
\description{
Take input methods and return extensions of analysis rds files.
}
