% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEAnalysis.R
\name{simul_methods}
\alias{simul_methods}
\title{Run Analysismethods}
\usage{
simul_methods(obj, output.dir, AnalysisMethods, para = list())
}
\arguments{
\item{obj}{Name and place of input files.}

\item{output.dir}{Result file save location.}

\item{AnalysisMethods}{DEmethods used for figures. Input as character vectors
(e.g. 'edgeR','edgeR.ql','edgeR.rb','DESeq.pd','DESeq2','voom.tmm','voom.qn','voom.sw','ROTS','BaySeq','BaySeq.qn,'PoissonSeq','SAMseq')}

\item{para}{A list parameter consists of multiple lists corresponding each method. Each method list contains multiple parameters to run each DE analysis method.}
}
\description{
Run Analysismethods
}
