% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationDataGeneration.R
\name{SyntheticDataSimulation}
\alias{SyntheticDataSimulation}
\title{Generate synthetic count data for analysis}
\usage{
SyntheticDataSimulation(simul.data, dataset, fixedfold = FALSE,
  samples.per.cond, n.var, n.diffexp, fraction.upregulated, dispType, mode,
  dataset.parameters)
}
\arguments{
\item{simul.data}{Characters indicating which dataset will be used for simulation data generation
"KIRC" for KIRC dataset.
"Bottomly" for Bottomly dataset.
"mKdB" for hybrid dataset combining mean of KIRC and dispersion of Bottomly datatset.
"mBdK" for hybrid dataset combining mean of Bottomly and dispersion of KIRC datatset.}

\item{dataset}{Characters specifying the file name of simulation dataset.}

\item{fixedfold}{A logical indicating whether this dataset is generated by fold changes with fixed values or random folds following exponential distribution.}

\item{samples.per.cond}{An integer indicating number of samples for each condition (e.g. 3).}

\item{n.var}{An integer indicating the number of total gene in the synthetic data.}

\item{n.diffexp}{An integer indicating number of generated DE genes in the synthetic data.}

\item{fraction.upregulated}{Proportion of upregulated DE genes in the synthetic data. (e.g. 0.5).}

\item{dispType}{Characters indicating how is the dispersion parameter assumed to be for each condition to make a synthetic data. Possible values are 'same' and 'different'.}

\item{mode}{Characters specifying test conditions used for simulation data generation.
"D" for basic simulation (not adding outliers).
"R" for adding 5% of random outlier.
"OS" for adding outlier sample to each sample group.
"DL" for decreasing KIRC simulation dispersion 22.5 times (similar to SEQC data dispersion) to compare with SEQC data.}

\item{dataset.parameters}{A list containing estimated mean and dispersion parameters and filtered count from original count dataset.}
}
\description{
Generate synthetic count data for analysis
}
